Add confirmable module to user.
In order to complete this, I had to add a migration for the columns
needed and add smtp settings to the application.rb. I also added some
lines to auto create records in the database (rake db:seed). Added a
method to sign in the test user before all of the feature tests.