Refactored most of the JS.
This is the initial stages of the new javascript structure I have decided on.
There are now namespaces containing functions that relate to the working of
the app, I plan on rearranging the file itself to make it a little more
presentable. I'm also considering moving the sub-namespaces into their own
files.