Fixed distraction time
The distraction time is now updated to include the time between editor
invocations, which really is since the last time a successful edit took
place in the database. This update also fixes the problem of word count
including html tags as words to be counted. It now does the right thing.