Completes goal based on time or word count
If there is a number in either field, it checks against the current
values to see if the goal is completed. It then updates the field
on the next 'put' ajax call made to the server. After this happens,
the distraction timer is stopped so that there is no longer any
distractions counting against a completed entry.