The beginning of context-based help.
This commit adds functionality in the way of context-based help. The idea
is to enter in a key/value pair into the Donuts.Help.HelpItems object. You
then take the key that you used and any add an id with the value of the key,
and a class of "help-item" and the browser will make a div appear relative
to the "help-item" element appear on hover and dissapear when not on hover.
The primitive styling is currently being done from within the javascript,
but it needs to have an entry in the CSS file.